requirements/testing.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/testing.txt requirements/testing.in
#
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
click==7.0
coverage==4.5.2
django-debug-toolbar==1.11
django-extensions==2.1.4
django-nose==1.4.6
django-redis==4.10.0
django-rest-framework==0.1.0
django==2.1.4
djangorestframework==3.9.0
filelock==3.0.10          # via tox
flake8==3.6.0
gunicorn==19.9.0
idna==2.8
kombu==4.4.0
mccabe==0.6.1             # via flake8
mock==2.0.0
nose==1.3.7               # via django-nose
pbr==5.1.3                # via mock
pip-tools==3.2.0
pluggy==0.9.0             # via tox
psycopg2-binary==2.7.7
psycopg2==2.7.7
py==1.8.0                 # via tox
pycodestyle==2.4.0        # via flake8
pyflakes==2.0.0           # via flake8
python-dateutil==2.7.5
pytz==2018.7
redis==3.2.0
requests==2.21.0
six==1.12.0
sqlparse==0.2.4           # via django-debug-toolbar
toml==0.10.0              # via tox
tox==3.6.1
urllib3==1.24.1
vine==1.2.0
virtualenv==16.4.3        # via tox
